Change model attribute and move models
As simulation instead of scenario was used, the models couldn't
stay in base_models.py because of circular import of Simulation
model class.
