+++ app/projects/models/__init__.py
@@ -1,6 +1,8 @@
 from projects.models.base_models import *
 from projects.models.simulation_models import (
     Simulation,
+    ParameterChangeTracker,
+    AssetChangeTracker,
     SensitivityAnalysis,
     get_project_sensitivity_analysis,
 )
