+++ app/projects/models/base_models.py
@@ -1,15 +1,13 @@
-import uuid
 import json
-from django.conf import settings
-from django.db import models
-from django.core.validators import MaxValueValidator, MinValueValidator
+import uuid
 from datetime import timedelta
+
+import oemof.thermal.compression_heatpumps_and_chillers as cmpr_hp_chiller
+from django.conf import settings
+from django.core.validators import MaxValueValidator, MinValueValidator
+from django.db import models
 from django.forms.models import model_to_dict
 from django.utils.translation import gettext_lazy as _
-from django.core.exceptions import ValidationError
-import oemof.thermal.compression_heatpumps_and_chillers as cmpr_hp_chiller
-
-from users.models import CustomUser
 from projects.constants import (
     ASSET_CATEGORY,
     ASSET_TYPE,
@@ -24,9 +22,8 @@ from projects.constants import (
     TRUE_FALSE_CHOICES,
     BOOL_CHOICES,
     USER_RATING,
-    PARAM_CATEGORY,
-    PARAM_TYPE,
 )
+from users.models import CustomUser
 
 
 class Feedback(models.Model):
@@ -528,34 +525,6 @@ class Asset(TopologyNode):
 
     def is_input_timeseries_empty(self):
         return self.input_timeseries == ""
-
-
-class ParameterInput(models.Model):
-    name = models.CharField(max_length=60, null=False, blank=False)
-    scenario = models.ForeignKey(
-        Scenario, on_delete=models.CASCADE, null=False, blank=False
-    )
-    old_value = models.TextField(null=True, blank=False)
-    new_value = models.TextField(null=True, blank=False)
-    parameter_category = models.CharField(max_length=20, choices=PARAM_CATEGORY)
-    parameter_type = models.CharField(
-        null=True, blank=True, max_length=20, choices=PARAM_TYPE
-    )
-    asset = models.ForeignKey(Asset, on_delete=models.CASCADE, null=True, blank=True)
-
-    def __str__(self):
-        return f"{self.name} [{self.old_value}, {self.new_value}]"
-
-
-class AssetChangeTracker(models.Model):
-    name = models.CharField(max_length=60, null=False, blank=False)
-    scenario = models.ForeignKey(
-        Scenario, on_delete=models.CASCADE, null=False, blank=False
-    )
-    asset_dto = models.TextField(null=True, blank=True)
-    action = models.SmallIntegerField(
-        null=False, blank=False, choices=((0, "delete"), (0, "create"))
-    )
 
 
 class COPCalculator(models.Model):
